agents/language_agent/requirements-language.txt
fastapi
pydantic
python-dotenv
jinja2
google-generativeai
